+++ src/Paginator/PaginatorInterface.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Data\Paginator;
 
+use InvalidArgumentException;
 use LogicException;
 use Yiisoft\Data\Reader\FilterInterface;
 use Yiisoft\Data\Reader\ReadableDataInterface;
@@ -34,6 +35,8 @@ interface PaginatorInterface extends ReadableDataInterface
      *
      * @param PageToken|null $token Page token. `Null` if current page is first.
      *
+     * @throws InvalidPageException If page token is incorrect.
+     *
      * @return static New instance.
      *
      * @see PageToken
@@ -45,9 +48,11 @@ interface PaginatorInterface extends ReadableDataInterface
      *
      * @param int $pageSize Maximum number of items per page.
      *
-     * @throws PaginatorException If page size is incorrect.
+     * @throws InvalidArgumentException If page size is not a positive number.
      *
      * @return static New instance.
+     *
+     * @psalm-param positive-int $pageSize
      */
     public function withPageSize(int $pageSize): static;
 
@@ -145,8 +150,6 @@ interface PaginatorInterface extends ReadableDataInterface
 
     /**
      * Get whether the current page is the last one.
-     *
-     * @throws PageNotFoundException If the page specified isn't found.
      *
      * @return bool Whether the current page is the last one.
      */