+++ src/Reader/DataReaderInterface.php
@@ -19,11 +19,13 @@ use IteratorAggregate;
  * @template TKey as array-key
  * @template TValue as array|object
  *
- * @extends ReadableDataInterface<TKey, TValue>
+ * @extends LimitableDataInterface<TKey, TValue>
+ * @extends OffsetableDataInterface<TKey, TValue>
+ * @extends SortableDataInterface<TKey, TValue>
+ * @extends FilterableDataInterface<TKey, TValue>
  * @extends IteratorAggregate<TKey, TValue>
  */
 interface DataReaderInterface extends
-    ReadableDataInterface,
     LimitableDataInterface,
     OffsetableDataInterface,
     CountableDataInterface,