+++ src/Reader/OffsetableDataInterface.php
@@ -6,8 +6,13 @@ namespace Yiisoft\Data\Reader;
 
 /**
  * Data that could be read from Nth item by skipping items from the beginning.
+ *
+ * @template TKey as array-key
+ * @template TValue as array|object
+ *
+ * @extends ReadableDataInterface<TKey, TValue>
  */
-interface OffsetableDataInterface
+interface OffsetableDataInterface extends ReadableDataInterface
 {
     /**
      * Get a new instance with an offset set.