+++ src/Reader/SortableDataInterface.php
@@ -7,8 +7,13 @@ namespace Yiisoft\Data\Reader;
 /**
  * A data set that could be sorted.
  * The sorting set may be used by a data reader that supports sorting.
+ *
+ * @template TKey as array-key
+ * @template TValue as array|object
+ *
+ * @extends ReadableDataInterface<TKey, TValue>
  */
-interface SortableDataInterface
+interface SortableDataInterface extends ReadableDataInterface
 {
     /**
      * Get a new instance with a sorting set.